feat: workshop — offline product entry + decimal price input
Cache the catalog locally so line entry works offline, and accept
fractional prices.

- Add a sembast `catalog` store + CatalogRepo; the add-line form searches
  the cache, never the network. WorkshopHome best-effort re-syncs from the
  API on open and via a manual refresh button; offline keeps the last cache.
- Switch money from whole-rupiah int to double (literal rupiah, <=2 decimals,
  DECIMAL(19,2)). MoneyInputFormatter replaces ThousandsInputFormatter:
  comma thousands + dot decimal, matching the console money-input, so the
  price field accepts a decimal point. Old int-valued stored jobs read back
  via (num).toDouble().

+++ packages/workshop/lib/component/margin_pill.dart
@@ -13,8 +13,8 @@ class CostMarginLine extends StatelessWidget {
     required this.priceMinor,
   });
 
-  final int costMinor;
-  final int priceMinor;
+  final double costMinor;
+  final double priceMinor;
 
   @override
   Widget build(BuildContext context) {
